refactor: remove unnecessary init files from groapiclient package
