+++ api/client/__init__.py
@@ -1,2 +1,2 @@
-from groapiclient import GroClient as Client
-from groapiclient import lib
+from groapiclient.client import GroClient as Client
+from groapiclient import lib, cfg
