--- groapiclient/batch_client/__init__.py
@@ -1,2 +1,0 @@
-from groapiclient import GroClient as BatchClient
-from groapiclient.client import BatchError
