--- groapiclient/crop_model/__init__.py
@@ -1,1 +1,0 @@
-from gro import GroClient as CropModel
