--- groapiclient/gro_client/__init__.py
@@ -1,2 +1,0 @@
-from groapiclient import GroClient
-from groapiclient.client import APIError
